app/requirements.txt
flask
flask-restful
requests
python-firebase